Extract maybeConvertHeic to utilities directory
Move the HEIC-to-JPEG conversion helper from image-drop-zone.tsx to
src/types/canary/utilities/maybe-convert-heic.ts so it can be reused
across components.
