+++ src/types/canary/utilities/index.ts
@@ -4,3 +4,4 @@ export * from './date-time';
 export * from './image-field-config';
 export * from './get-initials';
 export * from './get-cropped-image';
+export * from './maybe-convert-heic';
